requirements.txt
pyro-ppl
torch
matplotlib